move AverageSGD Tracker to utils/parameter
